Bringing main prod version into my branch

+++ backend/friendship/views.py
@@ -1,60 +1,19 @@
-from django.http import HttpResponse
 from django.db.models import Q
-from rest_framework import generics, status
+from rest_framework import status
 from rest_framework.response import Response
-from rest_framework.views import APIView
 from .models import FriendShip, FriendRequest, RelationShipStatus, BlockList
-from .serializers import FriendshipSerializer, FriendRequestSerializer, BlockListSerializer
+from .serializers import FriendRequestSerializer, BlockListSerializer
 from users.models import MyUser
-from conversations.models import Message
-
-from rest_framework.permissions import IsAuthenticated, AllowAny
 from rest_framework.decorators import api_view
 from conversations.models  import Channel
 from users.serializers import PublicUserSerializer
-# # Create your views here.
 
-class FriendRequestList(generics.ListAPIView):
-	permission_classes = [AllowAny]
-	serializer_class = FriendshipSerializer
-	def get_sentedRequest(self):
-		user_id = self.kwargs['user_id']
-		# Get all pending friend requests related to the user
-		return FriendShip.objects
-
-import random
-
-# Jhin and Jinx quotes
-jhin_quotes = [
-    "Art requires a certain... cruelty.",
-    "In carnage, I bloom, like a flower in the dawn.",
-    "I will make you beautiful.",
-    "Behind every mask... is another mask."
-]
-
-jinx_quotes = [
-    "Pow! Ha ha ha!",
-    "I'm crazy! Got a doctor's note.",
-    "Rules are made to be broken... like buildings! Or people!",
-    "Time to put on my dancing shoes!"
-]
-
-# Randomize between Jhin or Jinx, then select a random quote
-def random_quote():
-    character = random.choice(["Jhin", "Jinx"])
-    if character == "Jhin":
-        quote = random.choice(jhin_quotes)
-    else:
-        quote = random.choice(jinx_quotes)
-
-    return f"{character} says: \"{quote}\""
-##################################################################
-
-@api_view(['GET'])
+@api_view(['POST'])
 def AcceptFriendRequest(request, id):
 	try:
-		print('hmmm--')
 		friend_request = FriendRequest.objects.get(id=id)
+		if friend_request.receiver != request.user:
+			return Response({'message': 'friend request not related to this user'}, status=401)
 		list1, create = BlockList.objects.get_or_create(user=friend_request.sender)
 		check1 = list1.block_users.filter(id=friend_request.receiver.id).exists()
 		list2, create = BlockList.objects.get_or_create(user=friend_request.receiver)
@@ -74,16 +33,20 @@ def AcceptFriendRequest(request, id):
 			channel = Channel.objects.create()
 			channel.users.add(friend_request.sender)
 			channel.users.add(friend_request.receiver)
-			# message = Message.objects.create(sender=friend_request.sender, id_channel_fk=channel, content=random_quote())
 		return Response({'message': 'Accept request sent'}, status=status.HTTP_200_OK)
 	except Exception as e:
 		return Response({'Error': 'Something went wrong?' + str(e)}, status=status.HTTP_400_BAD_REQUEST)
 
-@api_view(['GET'])
+@api_view(['POST'])
 def BlockUser(request, unique_id):
 	try:
 		blocked_user = MyUser.objects.get(unique_id=unique_id)
 		myuser = request.user
+		# first let check if i m already blocked
+		blocked_user_list,create  = BlockList.objects.get_or_create(user=blocked_user)
+		check  = blocked_user_list.block_users.filter(id=myuser.id).exists()
+		if check:
+			return Response({'message': 'User '+unique_id+' already blocking you', 'status': 'unblocked'}, status=status.HTTP_200_OK)
 		list, create = BlockList.objects.get_or_create(user=myuser)
 		check = list.block_users.filter(id=blocked_user.id).exists()
 		if (not check):
@@ -97,11 +60,12 @@ def BlockUser(request, unique_id):
 				friendship.first().delete()
 			if len(f_request) > 0:
 				f_request.first().delete()
+
 		return Response({'message': 'User '+unique_id+' in the Block List', 'status': 'blocked'}, status=status.HTTP_200_OK)
 	except:
 		return Response({'Error': 'Something went wrong?'}, status=status.HTTP_400_BAD_REQUEST)
 
-@api_view(['GET'])
+@api_view(['POST'])
 def UnBlockUser(request, unique_id):
 	try:
 		blocked_user = MyUser.objects.get(unique_id=unique_id)
@@ -113,7 +77,7 @@ def UnBlockUser(request, unique_id):
 		return Response({'Error': 'Something went wrong?'}, status=status.HTTP_400_BAD_REQUEST)
 
 
-@api_view(['GET'])
+@api_view(['POST'])
 def UnFriendUser(request, unique_id):
 	try:
 		# intented solution you can postman this even whne you are blocked (my logic my rules contact hsaktiwy )
@@ -209,27 +173,21 @@ def FriendRequestStatus(request, uuid):
 	except:
 		return Response({'Error': 'Something went wrong?'}, status=status.HTTP_400_BAD_REQUEST)
 
-@api_view(['GET'])
+@api_view(['DELETE'])
 def CancelFriendRequest(request, id):
 	try:
 		myuser = request.user
 		friendrequest = FriendRequest.objects.filter(id=id)
 		if len(friendrequest) > 0:
+			if friendrequest.first().sender != myuser and friendrequest.first().receiver != myuser:
+				return Response({'message': 'friend request not related to this user'}, status=401)
 			friendrequest.first().delete()
-		return Response({'status': 'Done'}, status=status.HTTP_200_OK)
+			return Response({'status': 'Done'}, status=status.HTTP_200_OK)
+		else:
+			return Response({'message': 'Notification not found'}, status=404)
 	except:
 		return Response({'Error': 'Something went wrong?'}, status=status.HTTP_400_BAD_REQUEST)
-# class FriendRequestSentList(generics.ListAPIView):
-#     serializer_class = FriendRequestSerializer
-#     def get_queryset(self):
-#         user=self.request.user
-#         return FriendRequest.objects.filter(sender=user, status='pending')
 
-# class FriendRequestReceivedList(generics.ListAPIView):
-#     serializer_class = FriendRequestSerializer
-#     def get_queryset(self):
-#         user=self.request.user
-#         return FriendRequest.objects.filter(receiver=user, status='pending')
 @api_view(['GET'])
 def FriendRequestReceivedList(request):
 	try:
@@ -273,16 +231,9 @@ def FriendsList(request):
 		return Response(serialized_data.data, status=status.HTTP_200_OK)
 	except Exception as e:
 		return Response({'Error':str(e)}, status=status.HTTP_400_BAD_REQUEST)
-	# try:
-	# 	user = request.user
-	# 	FriendRequests = FriendRequest.objects.filter(sender=user)
-
-	# except:
-	# 	return Response({'Error': 'Something went wrong?'}, status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(['GET'])
 def GetBlockList(request):
-	print("allo w9")
 	try:
 		user= request.user
 		block_list = BlockList.objects.filter(user=user)
